Add tests for annotationSearch
